### 2021-12-17 #4000 @whedon
[PRE REVIEW]: liionpack: A Python package for simulating packs of batteries with PyBaMM
**Submitting author:** @TomTranter (<a href="http://orcid.org/0000-0003-4721-5941">Thomas Tranter</a>)
**Repository:** <a href="https://github.com/pybamm-team/liionpack" target ="_blank">https://github.com/pybamm-team/liionpack</a>
**Version:** v1.0.0
