### 2021-12-18 #4000 @jgostick
@tomtranter was my postdoc for 2 years from 2018-2019 and we have a slew of papers togehter.  If this is not a conflict of interest, the I'm OK to do it.  Just checking with you first.
