+++ paper/paper.bib
@@ -33,7 +33,8 @@
   number={1},
   pages={1--36},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12532-018-0139-4}
 }
 
 @article{pybamm,
@@ -43,7 +44,8 @@
   volume={9},
   number={1},
   year={2021},
-  publisher={Ubiquity Press}
+  publisher={Ubiquity Press},
+  doi={10.5334/jors.309}
 }
 
 @book{ltspice,
@@ -61,7 +63,8 @@
   number={6},
   pages={504--509},
   year={1975},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TCS.1975.1084079}
 }
 
 @article{jellyroll,
@@ -72,7 +75,8 @@
   number={11},
   pages={110538},
   year={2020},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1149/1945-7111/aba44b}
 }
 
 @article{Chen2020,
@@ -83,7 +87,8 @@
   number={8},
   pages={080534},
   year={2020},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1149/1945-7111/ab9050}
 }
 
 @article{Marquis2020,
@@ -94,5 +99,6 @@
   number={14},
   pages={140513},
   year={2020},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1149/1945-7111/abbce4}
 }
