### 2022-01-06 #4000 @kthyng
Hi @TomTranter — sorry but while I can't quite remember the rule of thumb for how long a conflict between postdoc advisor/postdoc is, it is at least 5 years and this is not a gray area. I see at least 4 other editors that may be appropriate for this submission but they are all currently full. Hopefully it won't be too long to get some availability! Thanks for your patience.
