### 2022-01-06 548115a
change the model timescale to a scalar
