### 2022-01-07 b68099b
Use a typical timescale if contains input
