### 2022-01-07 9c61ee2
don't use vectorised circuit solve
