### 2022-01-07 b864bde
Pass back none for initial concs and don't update if method depends on input
