### 2022-01-08 #4000 @danielskatz
fwiw, JOSS COI guidelines are in https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy
