### 2022-01-10 #4000 @timtroendle
:wave: @mefuller, @yangbai90 & @EricaEgg, would any of you be willing to review this submission for JOSS?
