### 2022-01-11 8040754
Use fixed timescale in all simulations and make new thermal_external simulation
