+++ liionpack/utils.py
@@ -47,7 +47,7 @@ def _convert_dict_to_list_of_dict(inputs_dict):
     return dicts
 
 
-def build_inputs_dict(I_batt, inputs):
+def build_inputs_dict(I_batt, inputs, external_variables):
     """
     Function to convert inputs and external_variable arrays to list of dicts
     As expected by the casadi solver. These are then converted back for mapped
@@ -60,6 +60,9 @@ def build_inputs_dict(I_batt, inputs):
         inputs (dict):
             A dictionary with key of each input and value an array of input
             values for each battery.
+        external_varaibles (dict):
+            A dictionary with key of each external variables and value an array
+            of variable values for each battery.
 
     Returns:
         list:
@@ -68,7 +71,11 @@ def build_inputs_dict(I_batt, inputs):
 
 
     """
-    inputs_dict = {"Current function [A]": I_batt}
+    inputs_dict = {}
+    if external_variables is not None:
+        inputs_dict.update(external_variables)
+    current_dict = {"Current function [A]": I_batt}
+    inputs_dict.update(current_dict)
     if inputs is not None:
         inputs_dict.update(inputs)
     inputs_dict = _convert_dict_to_list_of_dict(inputs_dict)
