+++ liionpack/__init__.py
@@ -7,6 +7,7 @@ read a netlist.
 """
 from .simulations import basic_simulation
 from .simulations import thermal_simulation
+from .simulations import thermal_external
 from .utils import interp_current
 from .utils import build_inputs_dict
 from .utils import add_events_to_model
