### 2022-01-11 fd3fea3
continue adding external_variable
