### 2022-01-12 #4051 @whedon
[REVIEW]: liionpack: A Python package for simulating packs of batteries with PyBaMM
**Submitting author:** @TomTranter (<a href="http://orcid.org/0000-0003-4721-5941">Thomas Tranter</a>)
**Repository:** <a href="https://github.com/pybamm-team/liionpack" target ="_blank">https://github.com/pybamm-team/liionpack</a>
**Version:** v0.3

- [x] #4141
