### 2022-01-12 #4051 @TomTranter
@timtroendle I fixed those DOIs in a review branch that I haven't merged yet. If there are any other necessary code changes I will also make them here https://github.com/pybamm-team/liionpack/tree/JOSS-review
