+++ liionpack/solvers.py
@@ -44,9 +44,7 @@ class generic_actor:
                 or (type(initial_soc) is list and len(initial_soc) == 1)
                 or (type(initial_soc) is np.ndarray and len(initial_soc) == 1)
             ):
-                _, _ = lp.update_init_conc(parameter_values,
-                                           initial_soc,
-                                           update=True)
+                _, _ = lp.update_init_conc(parameter_values, initial_soc, update=True)
             else:
                 lp.logger.warning(
                     "Using a list or an array of initial_soc "
@@ -201,9 +199,9 @@ class generic_manager:
         # Handle the inputs
         self.inputs = inputs
         self.external_variables = external_variables
-        self.inputs_dict = lp.build_inputs_dict(self.shm_i_app[0, :],
-                                                self.inputs,
-                                                self.external_variables)
+        self.inputs_dict = lp.build_inputs_dict(
+            self.shm_i_app[0, :], self.inputs, self.external_variables
+        )
         # Solver specific setup
         self.setup_actors(nproc, self.inputs_dict, initial_soc)
         # Get the initial state of the system
@@ -248,9 +246,9 @@ class generic_manager:
                     # for the next step
                     I_app = I_batt[:] * -1
                     self.shm_i_app[step + 1, :] = I_app
-                    self.inputs_dict = lp.build_inputs_dict(I_app,
-                                                            self.inputs,
-                                                            self.external_variables)                
+                    self.inputs_dict = lp.build_inputs_dict(
+                        I_app, self.inputs, self.external_variables
+                    )
                 # 06 Check if voltage limits are reached and terminate
                 if np.any(temp_v < v_cut_lower):
                     lp.logger.warning("Low voltage limit reached")
