### 2022-01-13 7bb3644
revert solver_utils solve method
