### 2022-01-13 649649c
Add step methods to generic_manager
