+++ liionpack/solvers.py
@@ -54,7 +54,7 @@ class generic_actor:
         if sim_func is None:
             self.simulation = lp.basic_simulation(self.parameter_values)
         else:
-            self.simulation = sim_func(parameter_values)
+            self.simulation = sim_func(self.parameter_values)
 
         # Set up integrator
         casadi_objs = cco(
@@ -285,6 +285,163 @@ class generic_manager:
             "Time per step " + str(np.around((toc - sim_start_time) / Nsteps, 3)) + "s"
         )
         return self.all_output
+
+    def step_solve(
+        self,
+        netlist,
+        sim_func,
+        parameter_values,
+        experiment,
+        inputs,
+        external_variables,
+        output_variables,
+        initial_soc,
+        nproc,
+    ):
+        self.netlist = netlist
+        self.sim_func = sim_func
+
+        self.parameter_values = parameter_values
+        self.check_current_function()
+        # Get netlist indices for resistors, voltage sources, current sources
+        self.Ri_map = netlist["desc"].str.find("Ri") > -1
+        self.V_map = netlist["desc"].str.find("V") > -1
+        self.I_map = netlist["desc"].str.find("I") > -1
+        self.Terminal_Node = np.array(netlist[self.I_map].node1)
+        self.Nspm = np.sum(self.V_map)
+
+        self.split_models(self.Nspm, nproc)
+
+        # Generate the protocol from the supplied experiment
+        self.protocol = lp.generate_protocol_from_experiment(experiment, flatten=True)
+        self.dt = experiment.period
+        self.Nsteps = len(self.protocol)
+        netlist.loc[self.I_map, ("value")] = self.protocol[0]
+        # Solve the circuit to initialise the electrochemical models
+        V_node, I_batt = lp.solve_circuit(netlist)
+
+        # The simulation output variables calculated at each step for each battery
+        # Must be a 0D variable i.e. battery wide volume average - or X-averaged for
+        # 1D model
+        self.variable_names = [
+            "Terminal voltage [V]",
+            "Measured battery open circuit voltage [V]",
+        ]
+        if output_variables is not None:
+            for out in output_variables:
+                if out not in self.variable_names:
+                    self.variable_names.append(out)
+            # variable_names = variable_names + output_variables
+        self.Nvar = len(self.variable_names)
+
+        # Storage variables for simulation data
+        self.shm_i_app = np.zeros([self.Nsteps, self.Nspm], dtype=float)
+        self.shm_Ri = np.zeros([self.Nsteps, self.Nspm], dtype=float)
+        self.output = np.zeros([self.Nvar, self.Nsteps, self.Nspm], dtype=float)
+
+        # Initialize currents in battery models
+        self.shm_i_app[0, :] = I_batt * -1
+
+        # Step forward in time
+        self.V_terminal = []
+        self.record_times = []
+
+        self.v_cut_lower = parameter_values["Lower voltage cut-off [V]"]
+        self.v_cut_higher = parameter_values["Upper voltage cut-off [V]"]
+
+        # Handle the inputs
+        self.inputs = inputs
+        self.inputs_dict = lp.build_inputs_dict(
+            self.shm_i_app[0, :], self.inputs, external_variables
+        )
+        # Solver specific setup
+        self.setup_actors(nproc, self.inputs_dict, initial_soc)
+        # Get the initial state of the system
+        self.evaluate_actors()
+
+    def _step_solve_step(self, external_variables):
+        tic = ticker.time()
+        # Do stepping
+        lp.logger.notice("Starting step solve")
+        vlims_ok = True
+        with tqdm(total=self.Nsteps, desc="Stepping simulation") as pbar:
+            step = 0
[…]
