### 2022-01-13 e1e059d
remove heat transfer input from thermal_external simulation as it doesn't get used
