### 2022-01-13 bc8cd7d
Add example of how to use the step method
