### 2022-01-19 083669d
Add power loss to netlist
