### 2022-01-25 30e3e2c
Update test_on_push.yml
install develop branch of pybamm and don't do conda for now
