+++ .github/workflows/test_on_push.yml
@@ -47,6 +47,10 @@ jobs:
           texlive-latex-extra \
           texlive-pictures
         sudo apt-get install ghostscript
+    - name: Install specific version of PyBaMM and liionpack
+      run: |
+        pip install wheel
+        python -m pip install git+https://github.com/pybamm-team/PyBaMM@develop
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
@@ -58,15 +62,6 @@ jobs:
       run: |
         python -m pip install coverage
         coverage run -m unittest -v
-    - name: Install dependencies in conda environment
-      run: |
-        $CONDA/bin/conda env update --file environment.yml --name base
-    - name: Install liionpack in conda environment for ray
-      run: |
-        $CONDA/bin/python setup.py install
-    - name: Run tests in the conda environment 
-      run: |
-        $CONDA/bin/python -m unittest -v
     - name: Build docs
       run: |
         python -m pip install -r docs/requirements.txt
