### 2022-01-27 318e5fe
replace solve with the step_solve method but rename that and give option to setup_only in order to step outside solver. Solve is essentially step solve with no external variable variation.
