+++ liionpack/solvers.py
@@ -132,6 +132,161 @@ class generic_manager:
     ):
         pass
 
+    # def solve(
+    #     self,
+    #     netlist,
+    #     sim_func,
+    #     parameter_values,
+    #     experiment,
+    #     inputs,
+    #     external_variables,
+    #     output_variables,
+    #     initial_soc,
+    #     nproc,
+    # ):
+    #     self.netlist = netlist
+    #     self.sim_func = sim_func
+
+    #     self.parameter_values = parameter_values
+    #     self.check_current_function()
+    #     # Get netlist indices for resistors, voltage sources, current sources
+    #     Ri_map = netlist["desc"].str.find("Ri") > -1
+    #     V_map = netlist["desc"].str.find("V") > -1
+    #     I_map = netlist["desc"].str.find("I") > -1
+    #     Terminal_Node = np.array(netlist[I_map].node1)
+    #     Nspm = np.sum(V_map)
+
+    #     self.split_models(Nspm, nproc)
+
+    #     # Generate the protocol from the supplied experiment
+    #     protocol = lp.generate_protocol_from_experiment(experiment, flatten=True)
+    #     self.dt = experiment.period
+    #     Nsteps = len(protocol)
+    #     netlist.loc[I_map, ("value")] = protocol[0]
+    #     # Solve the circuit to initialise the electrochemical models
+    #     V_node, I_batt = lp.solve_circuit(netlist)
+
+    #     # The simulation output variables calculated at each step for each battery
+    #     # Must be a 0D variable i.e. battery wide volume average - or X-averaged for
+    #     # 1D model
+    #     self.variable_names = [
+    #         "Terminal voltage [V]",
+    #         "Measured battery open circuit voltage [V]",
+    #     ]
+    #     if output_variables is not None:
+    #         for out in output_variables:
+    #             if out not in self.variable_names:
+    #                 self.variable_names.append(out)
+    #         # variable_names = variable_names + output_variables
+    #     Nvar = len(self.variable_names)
+
+    #     # Storage variables for simulation data
+    #     self.shm_i_app = np.zeros([Nsteps, Nspm], dtype=float)
+    #     self.shm_Ri = np.zeros([Nsteps, Nspm], dtype=float)
+    #     self.output = np.zeros([Nvar, Nsteps, Nspm], dtype=float)
+
+    #     # Initialize currents in battery models
+    #     self.shm_i_app[0, :] = I_batt * -1
+
+    #     # Step forward in time
+    #     self.timestep = 0
+    #     V_terminal = []
+    #     record_times = []
+
+    #     v_cut_lower = parameter_values["Lower voltage cut-off [V]"]
+    #     v_cut_higher = parameter_values["Upper voltage cut-off [V]"]
+
+    #     # Handle the inputs
+    #     self.inputs = inputs
+    #     self.external_variables = external_variables
+    #     self.inputs_dict = lp.build_inputs_dict(
+    #         self.shm_i_app[0, :], self.inputs, self.external_variables
+    #     )
+    #     # Solver specific setup
+    #     self.setup_actors(nproc, self.inputs_dict, initial_soc)
+    #     # Get the initial state of the system
+    #     self.evaluate_actors()
+    #     sim_start_time = ticker.time()
+    #     lp.logger.notice("Starting step solve")
+    #     with tqdm(total=Nsteps, desc="Stepping simulation") as pbar:
+    #         step = 0
+    #         while step < Nsteps:
+    #             # 01 Calculate whether resting or restarting
+    #             self.resting = (
+    #                 step > 0 and protocol[step] == 0.0 and protocol[step - 1] == 0.0
+    #             )
+    #             self.restarting = (
[…]
