### 2022-01-27 09927d3
Change error to warning on input for initial SoC and add test
