+++ tests/unit/test_sim_utils.py
@@ -22,6 +22,16 @@ class sim_utilsTest(unittest.TestCase):
         with self.assertRaises(ValueError):
             lp.update_init_conc(self.param, SoC=10.0)
 
+    def test_input_logger_warning(self):
+        param = pybamm.ParameterValues(chemistry="Chen2020")
+        neg_conc = 0.0
+        param.update({
+            "Electrode height [m]": "[input]",
+            "Initial concentration in negative electrode [mol.m-3]": neg_conc
+            }, check_already_exists=False)
+        lp.update_init_conc(param, SoC=0.5, update=True)
+        a = param["Initial concentration in negative electrode [mol.m-3]"]
+        assert a == neg_conc
 
 if __name__ == "__main__":
     unittest.main()
