+++ tests/unit/test_sim_utils.py
@@ -25,13 +25,17 @@ class sim_utilsTest(unittest.TestCase):
     def test_input_logger_warning(self):
         param = pybamm.ParameterValues(chemistry="Chen2020")
         neg_conc = 0.0
-        param.update({
-            "Electrode height [m]": "[input]",
-            "Initial concentration in negative electrode [mol.m-3]": neg_conc
-            }, check_already_exists=False)
+        param.update(
+            {
+                "Electrode height [m]": "[input]",
+                "Initial concentration in negative electrode [mol.m-3]": neg_conc,
+            },
+            check_already_exists=False,
+        )
         lp.update_init_conc(param, SoC=0.5, update=True)
         a = param["Initial concentration in negative electrode [mol.m-3]"]
         assert a == neg_conc
+
 
 if __name__ == "__main__":
     unittest.main()
