### 2022-01-27 d3afd0b
Add build_inputs test
