### 2022-02-02 pybamm-team/liionpack#128 @mefuller
JOSS Review: Paper
### Description

Paper needs references updated with DOIs where possible (see https://github.com/openjournals/joss-reviews/issues/4051#issuecomment-1010721400)

Also, please add some "literature review" or survey to identify what other software packages (if any) exist that provide similar capabilities and how these differ.

Thank you

### Motivation

_No response_

### Possible Implementation

_No response_

### Additional context

_No response_
