### 2022-02-03 dd63694
change lists to arrays
