### 2022-02-04 pybamm-team/liionpack#128 @TomTranter
@mefuller I have made a pull request to address these points but I can't assign you to review it. Shall I just merge it in?
