+++ examples/dask_solver.py
@@ -25,8 +25,7 @@ def main():
         period='10 seconds')
 
     # PyBaMM parameters
-    chemistry = pybamm.parameter_sets.Chen2020
-    parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+    parameter_values = pybamm.ParameterValues("Chen2020")
 
     # Solve pack using Dask solver
     output = lp.solve(
