+++ tests/integration/test_all_solvers.py
@@ -15,8 +15,7 @@ class TestSolvers(unittest.TestCase):
             Np=1, Ns=1, Rb=Rsmall, Rc=Rsmall, Ri=5e-2, V=4.0, I=1.0
         )
         # PyBaMM parameters
-        chemistry = pybamm.parameter_sets.Chen2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Chen2020")
         # Cycling experiment
         experiment = pybamm.Experiment(
             [("Discharge at 1 A for 100 s or until 3.3 V",)] * 1, period="10 s"
