### 2022-02-09 #4051 @EricaEgg
Great work!
I have completed my review. My only recommendation is that it was not clear in the documentation that LaTeX is required. Please add this to installation instructions.
