### 2022-02-10 bb293ec
Add latex instructions to readme
