### 2022-02-14 #4051 @timtroendle
@TomTranter, all looks good to me! There is however a lot of unused whitespace in the current paper version. May I suggest you pull Figure 1 above the "Algorithm" section heading? That will give you a much better text flow and it will also remove one (unused) page from the paper.

Apart from that, could you please:
