+++ .github/workflows/publish_pypi.yml
@@ -1,0 +1,65 @@
+name: Build and publish package to PyPI
+
+on:
+  workflow_dispatch:
+    inputs:
+      target:
+        description: 'Deployment target. Can be "pypi" or "testpypi"'
+        default: 'pypi'
+
+jobs:
+
+  build-pkg:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.8]
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: pip install wheel
+
+      - name: Build package
+        run: python setup.py sdist --formats=gztar bdist_wheel
+
+      - name: Upload package
+        uses: actions/upload-artifact@v2
+        with:
+          name: files
+          path: ./dist/
+          if-no-files-found: error
+
+
+  publish-pypi:
+    needs: build-pkg
+    runs-on: ubuntu-latest
+    steps:
+      - name: Download package
+        uses: actions/download-artifact@v2
+        with:
+          name: files
+          path: dist
+
+      - name: Publish on PyPI
+        if: github.event.inputs.target == 'pypi'
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          user: __token__
+          password: ${{ secrets.PYPI_TOKEN }}
+          packages_dir: dist/
+
+      - name: Publish on TestPyPI
+        if: github.event.inputs.target == 'testpypi'
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          user: __token__
+          password: ${{ secrets.TESTPYPI_TOKEN }}
+          packages_dir: dist/
+          repository_url: https://test.pypi.org/legacy/
