### 2022-02-18 #4051 @danielskatz
👋 @TomTranter - I'm the JOSS AEiC this week, who handles the remaining processing of your paper.

I see that the zenodo deposit has fairly different metadata than the paper.  The paper has 11 authors, the archive has 9. The paper authors don't have first names, only first initials, while the archive authors mostly have two names, except one who has three and one who only has one.  The titles of the paper and the archive are also quite different. Normally, we suggest that these be consistent.  Can you either make them consistent, or explain your thoughts on why they shouldn't be?

Also, the JOSS style is to use full names of authors, usually a first and last name, sometimes with one or more initials as well (as in the the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)).

I'm now going to proofread the paper and references, and may have further comments.
