### 2022-02-18 #4051 @danielskatz
I have the following suggested change in the paper: https://github.com/pybamm-team/liionpack/pull/142

In addition, can you define DAE when you first use it?
