### 2022-02-18 #4051 @TomTranter
@danielskatz the two authors who have not contributed code have contributed review of the paper and also supervision of the projects that the code was developed under. Is it ok to keep them on the paper? I will work to correct the other issues you raised. Apologies for the inconsistency it was my first time using zenodo
