### 2022-02-18 #4051 @danielskatz
If the other two authors contributed to the software, I would add them to the zenodo - I think it's intended to recognize contributors to the project, not just those who made commits in the repo.  But this is your decision.
